move crab_cfg into crab_constraints
